Initial society onboarding. Seperate routes
